content/canvas/test/test_2d.pattern.image.broken.html

<!DOCTYPE HTML>
<title>Canvas test: 2d.pattern.image.broken</title>
<script src="/MochiKit/packed.js"></script>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" href="/tests/SimpleTest/test.css">
<body>
<canvas id="c" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>
<script>

SimpleTest.waitForExplicitFinish();
MochiKit.DOM.addLoadEvent(function () {

var canvas = document.getElementById('c');
var ctx = canvas.getContext('2d');

var img = document.getElementById('broken.png');
todo(img.complete === false, "img.complete === false");
var _thrown = undefined; try {
  ctx.createPattern(img, 'repeat');
} catch (e) { _thrown = e }; todo(_thrown && _thrown.code == DOMException.INVALID_STATE_ERR, "should throw INVALID_STATE_ERR");

SimpleTest.finish();

});
</script>
<img src="image_broken.png" id="broken.png" class="resource">
